// src/types/graphql-types.d.ts
export type Maybe<T> = T | null
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /** A date string, such as 2007-12-03, compliant with the ISO 8601 standard for
   * representation of dates and times using the Gregorian calendar.
   */
  Date: any
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any
}
// ...
export type Directory = Node & {
  __typename?: "Directory"
  id: Scalars["ID"]
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
  sourceInstanceName?: Maybe<Scalars["String"]>
  absolutePath?: Maybe<Scalars["String"]>
  relativePath?: Maybe<Scalars["String"]>
  extension?: Maybe<Scalars["String"]>
  size?: Maybe<Scalars["Int"]>
  prettySize?: Maybe<Scalars["String"]>
  modifiedTime?: Maybe<Scalars["Date"]>
  accessTime?: Maybe<Scalars["Date"]>
  changeTime?: Maybe<Scalars["Date"]>
  birthTime?: Maybe<Scalars["Date"]>
  root?: Maybe<Scalars["String"]>
  dir?: Maybe<Scalars["String"]>
  base?: Maybe<Scalars["String"]>
  ext?: Maybe<Scalars["String"]>
  name?: Maybe<Scalars["String"]>
  relativeDirectory?: Maybe<Scalars["String"]>
  dev?: Maybe<Scalars["Int"]>
  mode?: Maybe<Scalars["Int"]>
  nlink?: Maybe<Scalars["Int"]>
  uid?: Maybe<Scalars["Int"]>
  gid?: Maybe<Scalars["Int"]>
  rdev?: Maybe<Scalars["Int"]>
  blksize?: Maybe<Scalars["Int"]>
  ino?: Maybe<Scalars["Int"]>
  blocks?: Maybe<Scalars["Int"]>
  atimeMs?: Maybe<Scalars["Float"]>
  mtimeMs?: Maybe<Scalars["Float"]>
  ctimeMs?: Maybe<Scalars["Float"]>
  birthtimeMs?: Maybe<Scalars["Float"]>
  atime?: Maybe<Scalars["Date"]>
  mtime?: Maybe<Scalars["Date"]>
  ctime?: Maybe<Scalars["Date"]>
  birthtime?: Maybe<Scalars["Date"]>
}
// ...
export type DirectoryConnection = {
  __typename?: "DirectoryConnection"
  totalCount: Scalars["Int"]
  edges: Array<DirectoryEdge>
  nodes: Array<Directory>
  pageInfo: PageInfo
  distinct: Array<Scalars["String"]>
  group: Array<DirectoryGroupConnection>
}
// ...
export type File = Node & {
  __typename?: "File"
  birthtime?: Maybe<Scalars["Date"]>
  birthtimeMs?: Maybe<Scalars["Float"]>
  sourceInstanceName?: Maybe<Scalars["String"]>
  absolutePath?: Maybe<Scalars["String"]>
  relativePath?: Maybe<Scalars["String"]>
  extension?: Maybe<Scalars["String"]>
  size?: Maybe<Scalars["Int"]>
  prettySize?: Maybe<Scalars["String"]>
  modifiedTime?: Maybe<Scalars["Date"]>
  accessTime?: Maybe<Scalars["Date"]>
  changeTime?: Maybe<Scalars["Date"]>
  birthTime?: Maybe<Scalars["Date"]>
  root?: Maybe<Scalars["String"]>
  dir?: Maybe<Scalars["String"]>
  base?: Maybe<Scalars["String"]>
  ext?: Maybe<Scalars["String"]>
  name?: Maybe<Scalars["String"]>
  relativeDirectory?: Maybe<Scalars["String"]>
  dev?: Maybe<Scalars["Int"]>
  mode?: Maybe<Scalars["Int"]>
  nlink?: Maybe<Scalars["Int"]>
  uid?: Maybe<Scalars["Int"]>
  gid?: Maybe<Scalars["Int"]>
  rdev?: Maybe<Scalars["Int"]>
  blksize?: Maybe<Scalars["Int"]>
  ino?: Maybe<Scalars["Int"]>
  blocks?: Maybe<Scalars["Int"]>
  atimeMs?: Maybe<Scalars["Float"]>
  mtimeMs?: Maybe<Scalars["Float"]>
  ctimeMs?: Maybe<Scalars["Float"]>
  atime?: Maybe<Scalars["Date"]>
  mtime?: Maybe<Scalars["Date"]>
  ctime?: Maybe<Scalars["Date"]>
  /** Copy file to static directory and return public url to it */
  publicURL?: Maybe<Scalars["String"]>
  id: Scalars["ID"]
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
  childImageSharp?: Maybe<ImageSharp>
}
// ...
export type FileConnection = {
  __typename?: "FileConnection"
  totalCount: Scalars["Int"]
  edges: Array<FileEdge>
  nodes: Array<File>
  pageInfo: PageInfo
  distinct: Array<Scalars["String"]>
  group: Array<FileGroupConnection>
}
// ...
export type ImageSharp = Node & {
  __typename?: "ImageSharp"
  id: Scalars["ID"]
  fixed?: Maybe<ImageSharpFixed>
  resolutions?: Maybe<ImageSharpResolutions>
  fluid?: Maybe<ImageSharpFluid>
  sizes?: Maybe<ImageSharpSizes>
  original?: Maybe<ImageSharpOriginal>
  resize?: Maybe<ImageSharpResize>
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
}
// ...
export type ImageSharpConnection = {
  __typename?: "ImageSharpConnection"
  totalCount: Scalars["Int"]
  edges: Array<ImageSharpEdge>
  nodes: Array<ImageSharp>
  pageInfo: PageInfo
  distinct: Array<Scalars["String"]>
  group: Array<ImageSharpGroupConnection>
}
// ...
export type Internal = {
  __typename?: "Internal"
  content?: Maybe<Scalars["String"]>
  contentDigest: Scalars["String"]
  description?: Maybe<Scalars["String"]>
  fieldOwners?: Maybe<Array<Maybe<Scalars["String"]>>>
  ignoreType?: Maybe<Scalars["Boolean"]>
  mediaType?: Maybe<Scalars["String"]>
  owner: Scalars["String"]
  type: Scalars["String"]
}
// ...
export type Node = {
  id: Scalars["ID"]
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
}
// ...
export type Query = {
  __typename?: "Query"
  imageSharp?: Maybe<ImageSharp>
  allImageSharp?: Maybe<ImageSharpConnection>
  file?: Maybe<File>
  allFile?: Maybe<FileConnection>
  sitePage?: Maybe<SitePage>
  allSitePage?: Maybe<SitePageConnection>
  sitePlugin?: Maybe<SitePlugin>
  allSitePlugin?: Maybe<SitePluginConnection>
  site?: Maybe<Site>
  allSite?: Maybe<SiteConnection>
  directory?: Maybe<Directory>
  allDirectory?: Maybe<DirectoryConnection>
}
// ...
export type Site = Node & {
  __typename?: "Site"
  id: Scalars["ID"]
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
  siteMetadata?: Maybe<SiteSiteMetadata>
  port?: Maybe<Scalars["Int"]>
  host?: Maybe<Scalars["String"]>
  polyfill?: Maybe<Scalars["Boolean"]>
  pathPrefix?: Maybe<Scalars["String"]>
  buildTime?: Maybe<Scalars["Date"]>
}

export type SiteBuildTimeArgs = {
  formatString?: Maybe<Scalars["String"]>
  fromNow?: Maybe<Scalars["Boolean"]>
  difference?: Maybe<Scalars["String"]>
  locale?: Maybe<Scalars["String"]>
}

export type SiteConnection = {
  __typename?: "SiteConnection"
  totalCount: Scalars["Int"]
  edges: Array<SiteEdge>
  nodes: Array<Site>
  pageInfo: PageInfo
  distinct: Array<Scalars["String"]>
  group: Array<SiteGroupConnection>
}
// ...
export type SitePage = Node & {
  __typename?: "SitePage"
  id: Scalars["ID"]
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
  path?: Maybe<Scalars["String"]>
  internalComponentName?: Maybe<Scalars["String"]>
  component?: Maybe<Scalars["String"]>
  componentChunkName?: Maybe<Scalars["String"]>
  isCreatedByStatefulCreatePages?: Maybe<Scalars["Boolean"]>
  pluginCreator?: Maybe<SitePlugin>
  pluginCreatorId?: Maybe<Scalars["String"]>
  componentPath?: Maybe<Scalars["String"]>
}

export type SitePageConnection = {
  __typename?: "SitePageConnection"
  totalCount: Scalars["Int"]
  edges: Array<SitePageEdge>
  nodes: Array<SitePage>
  pageInfo: PageInfo
  distinct: Array<Scalars["String"]>
  group: Array<SitePageGroupConnection>
}
// ...
export type SitePlugin = Node & {
  __typename?: "SitePlugin"
  id: Scalars["ID"]
  parent?: Maybe<Node>
  children: Array<Node>
  internal: Internal
  resolve?: Maybe<Scalars["String"]>
  name?: Maybe<Scalars["String"]>
  version?: Maybe<Scalars["String"]>
  pluginOptions?: Maybe<SitePluginPluginOptions>
  nodeAPIs?: Maybe<Array<Maybe<Scalars["String"]>>>
  ssrAPIs?: Maybe<Array<Maybe<Scalars["String"]>>>
  pluginFilepath?: Maybe<Scalars["String"]>
  packageJson?: Maybe<SitePluginPackageJson>
}

export type SitePluginConnection = {
  __typename?: "SitePluginConnection"
  totalCount: Scalars["Int"]
  edges: Array<SitePluginEdge>
  nodes: Array<SitePlugin>
  pageInfo: PageInfo
  distinct: Array<Scalars["String"]>
  group: Array<SitePluginGroupConnection>
}
// ...
export type SiteSiteMetadata = {
  __typename?: "SiteSiteMetadata"
  title?: Maybe<Scalars["String"]>
  description?: Maybe<Scalars["String"]>
  author?: Maybe<Scalars["String"]>
}
// ...
export type SiteMetadataQueryQuery = { __typename?: "Query" } & {
  site: Maybe<
    { __typename?: "Site" } & {
      siteMetadata: Maybe<
        { __typename?: "SiteSiteMetadata" } & Pick<
          SiteSiteMetadata,
          "title" | "description" | "author"
        >
      >
    }
  >
}